refactor package structure.
switched from requirements.txt to pyproject.toml.

switched from flat layout to src layout.

added convenience makefile to streamline installation with "common-denominator" build flags (and other useful instructions).
