+++ pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "llama_supercharged"
+version = "0.1.0"
+dependencies = [
+	"llama-cpp-python @ git+https://github.com/JamePeng/llama-cpp-python.git"
+]
+requires-python = ">=3.11" # (very) possibly lower.
+authors = [
+	{ name = "Azkali", email = "a.ffcc7@gmail.com" },
+	{ name = "NV4", email = "cattobyte@kyararanbii.xyz" },
+]
+description = "Supercharged interface library for running (and stringing) LLMs."
+readme = "README.md"
